Add files via upload
I didn't do a whole lot of troubleshooting with this code. It outputs a state space matrix, and needs a water usage profile to run. You can take T_in as just the indoor temperature which is from the HVAC Model. T_Win is the water inlet temperature, assume this is constant, maybe 16 C. I also have the emulated manufacturer controller to act as the "real system". Use the system state-space dynamics for the optimization code.